+++ src/Payable.php
@@ -3,12 +3,12 @@
 namespace Marshmallow\Payable;
 
 use Exception;
-use Marshmallow\Payable\Providers\Adyen;
 use Marshmallow\Payable\Providers\Mollie;
 use Marshmallow\Payable\Providers\Stripe;
 use Marshmallow\Payable\Models\PaymentType;
 use Marshmallow\Payable\Providers\Buckaroo;
 use Marshmallow\Payable\Providers\Provider;
+use Marshmallow\Payable\Providers\Worldline;
 use Marshmallow\Payable\Providers\MultiSafePay;
 
 class Payable
@@ -16,8 +16,8 @@ class Payable
     public const MOLLIE = 'MOLLIE';
     public const MULTI_SAFE_PAY = 'MULTI_SAFE_PAY';
     public const STRIPE = 'STRIPE';
-    public const ADYEN = 'ADYEN';
     public const BUCKAROO = 'BUCKAROO';
+    public const WORLDLINE = 'WORLDLINE';
 
     public function getProvider(PaymentType $paymentType): Provider
     {
@@ -34,13 +34,12 @@ class Payable
                 return new Stripe;
                 break;
 
-
-            case self::ADYEN:
-                return new Adyen;
-                break;
-
             case self::BUCKAROO:
                 return new Buckaroo;
+                break;
+
+            case self::WORLDLINE:
+                return new Worldline;
                 break;
         }