+++ src/PayableTest.php
@@ -29,18 +29,6 @@ class PayableTest
         return $cart->startPayment($payment_type, $test, $api_key);
     }
 
-    public function adyen($test = false, $api_key = null)
-    {
-        $cart = $this->getTestCart();
-        $payment_type = $this->getPaymentType(Payable::ADYEN);
-        if ($this->recurring) {
-            return $cart->startRecurringPayment($payment_type, $test, $api_key);
-        }
-        return $cart->startPayment($payment_type, $test, $api_key);
-    }
-
-
-
     public function buckaroo($test = true)
     {
         $cart = $this->getTestCart();